submit

+++ src/main/c/builtin-func.c
@@ -199,5 +199,5 @@ int __string_ord(char* str, int pos) {
 }
 
 int __array_size(char * arr) {
-    return *(((long long*)arr) - 1);
+    return *(((int*)arr) - 1);
 }
