+++ src/main/c/t.c
@@ -204,5 +204,5 @@ int __string_ord(char* str, int pos) {
 }
 
 int __array_size(char * arr) {
-    return *(((long long*)arr) - 1);
+    return *(((int*)arr) - 1);
 }
